Fix printf format specifier mismatch when REGEX_DEBUG

+++ regex.c
@@ -580,7 +580,7 @@ regmatch_iter(struct MatchingContext1 *c,
     c->lastpos = c->str;
 #ifdef REGEX_DEBUG
     if (verbose)
-	printf("Succeed: %s %d\n", c->str, c->lastpos - c->str);
+	printf("Succeed: %s %ld\n", c->str, (long)(c->lastpos - c->str));
 #endif
     YIELD(1, c, 7);
     return 0;
